Introduce `ClientExtensions` type
Instead of `Vec<ClientExtension>`, store the extension data as
a struct.  This is possible because past commits have removed
the need for this us to losslessly round-trip extension data.

This involves fewer allocations to construct the extensions for clients.

It eliminates repeated iteration of the vector to find specific extensions
when processing a `ClientHello` for servers.  It also reduces
the cost of detecting duplicate extensions.